svc/requirements.txt
# Web & Config
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.0

# LangChain and LLM dependencies
langchain==0.3.27
langchain-core>=0.3.72,<1.0.0
langchain-openai==0.3.35

# Gemini
google-generativeai==0.8.5
langchain-google-genai==2.0.10  # <3.0.0 for not using Python 3.10

# OpenAI
openai==2.3.0

# RAG / Vector DB
chromadb==0.5.3
langchain-chroma==0.1.4        # Updated to compatible version
langchain-text-splitters==0.3.11
pypdf==4.0.1
